+++ SRC/element/Frame/Other/Force/ForceBeamColumn3dThermal.cpp
@@ -3411,28 +3411,28 @@ ForceBeamColumn3dThermal::getInitialDeformations(Vector &v0)
     // section response -
     else if (strcmp(argv[0],"sectionX") == 0) {
       if (argc > 2) {
-	float sectionLoc = atof(argv[1]);
-	
-	double xi[maxNumSections];
-	double L = crdTransf->getInitialLength();
-	beamIntegr->getSectionLocations(numSections, L, xi);
-	
-	sectionLoc /= L;
-	
-	float minDistance = fabs(xi[0]-sectionLoc);
-	int sectionNum = 0;
-	for (int i = 1; i < numSections; i++) {
-	  if (fabs(xi[i]-sectionLoc) < minDistance) {
-	    minDistance = fabs(xi[i]-sectionLoc);
-	    sectionNum = i;
-	  }
-	}
-	
-	output.tag("GaussPointOutput");
-	output.attr("number",sectionNum+1);
-	output.attr("eta",xi[sectionNum]*L);
-	
-	theResponse = sections[sectionNum]->setResponse(&argv[2], argc-2, output);
+        double sectionLoc = atof(argv[1]);
+
+        double xi[maxNumSections];
+        double L = crdTransf->getInitialLength();
+        beamIntegr->getSectionLocations(numSections, L, xi);
+        
+        sectionLoc /= L;
+        
+        double minDistance = fabs(xi[0]-sectionLoc);
+        int sectionNum = 0;
+        for (int i = 1; i < numSections; i++) {
+          if (fabs(xi[i]-sectionLoc) < minDistance) {
+            minDistance = fabs(xi[i]-sectionLoc);
+            sectionNum = i;
+          }
+        }
+        
+        output.tag("GaussPointOutput");
+        output.attr("number",sectionNum+1);
+        output.attr("eta",xi[sectionNum]*L);
+        
+        theResponse = sections[sectionNum]->setResponse(&argv[2], argc-2, output);
       }
     }
